firefly: improve build.sh

+++ common/build.sh
@@ -4,15 +4,23 @@ CMD=`realpath $0`
 COMMON_DIR=`dirname $CMD`
 TOP_DIR=$(realpath $COMMON_DIR/../../..)
 BOARD_CONFIG=$TOP_DIR/device/rockchip/.BoardConfig.mk
+CFG_DIR=$TOP_DIR/device/rockchip
+ROCKDEV=$TOP_DIR/rockdev
 source $BOARD_CONFIG
 source $TOP_DIR/device/rockchip/common/Version.mk
+PARAMETER=$TOP_DIR/device/rockchip/$RK_TARGET_PRODUCT/$RK_PARAMETER
+SD_PARAMETER=$TOP_DIR/device/rockchip/$RK_TARGET_PRODUCT/$RK_SD_PARAMETER
 
 if [ ! -n "$1" ];then
 	echo "build all and save all as default"
 	BUILD_TARGET=allsave
 else
-	BUILD_TARGET="$1"
-	NEW_BOARD_CONFIG=$TOP_DIR/device/rockchip/$RK_TARGET_PRODUCT/$1
+	BUILD_TARGET=$1
+   	NEW_BOARD_CONFIG=$(find $CFG_DIR -name "$1")
+fi
+
+if [ ! -d "rockdev/pack" ];then
+	mkdir rockdev/pack
 fi
 
 usage()
@@ -21,6 +29,7 @@ usage()
 	echo "uboot              -build uboot"
 	echo "kernel             -build kernel"
 	echo "modules            -build kernel modules"
+	echo "extboot            -build extlinux boot.img, boot from EFI partition"
 	echo "rootfs             -build default rootfs, currently build buildroot as default"
 	echo "buildroot          -build buildroot rootfs"
 	echo "ramboot            -build ramboot image"
@@ -32,10 +41,40 @@ usage()
 	echo "all                -build uboot, kernel, rootfs, recovery image"
 	echo "cleanall           -clean uboot, kernel, rootfs, recovery"
 	echo "firmware           -pack all the image we need to boot up system"
-	echo "updateimg          -pack update image"
+	echo "sdupdateimg        -pack sdupdate image"
+	echo "updateimg [-p]     -pack update image; [-p] compress"
+	echo "sdbootimg [-p]     -pack sdboot image; [-p] compress"
 	echo "otapackage         -pack ab update otapackage image"
 	echo "save               -save images, patches, commands used to debug"
 	echo "default            -build all modules"
+	echo "BoardConfig Board  -select Board and it's BoardConfig.mk   "
+	exit
+}
+
+function build_extboot_image() {
+
+    build_kernel
+
+	BOOT=${TOP_DIR}/kernel/extboot.img
+	rm -rf ${BOOT}
+
+	echo -e "\e[36m Generate extLinuxBoot image start\e[0m"
+
+	# 100 Mb
+	mkfs.vfat -n "EFI" -S 512 -C ${BOOT} $((30 * 1024))
+
+	echo "label kernel-4.4" > temp.conf
+	echo "    kernel /Image" >> temp.conf
+	echo "    fdt /${RK_KERNEL_DTS}.dtb" >> temp.conf
+
+	mmd -i ${BOOT} ::/extlinux
+	mcopy -i ${BOOT} -s temp.conf ::/extlinux/extlinux.conf
+	mcopy -i ${BOOT} -s ${TOP_DIR}/kernel/arch/${RK_ARCH}/boot/dts/rockchip/${RK_KERNEL_DTS}.dtb ::
+	mcopy -i ${BOOT} -s ${TOP_DIR}/kernel/arch/${RK_ARCH}/boot/Image ::
+
+	rm temp.conf
+
+	echo -e "\e[36m Generate extLinux Boot image : ${BOOT} success! \e[0m"
 }
 
 function build_uboot(){
@@ -218,7 +257,80 @@ function build_firmware(){
 	fi
 }
 
+
+function gen_file_name() {
+	day=$(date +%Y%m%d)
+	time=$(date +%H%M)
+
+	typeset -u board
+	board=$(basename $(readlink ${BOARD_CONFIG}))
+	board=${board%%.MK}
+		
+	rootfs=$(ls -l rockdev/ | grep rootfs.img | awk -F '/' '{print $(NF)}'|awk -F '_' '{print $2}')
+
+	board=${board}${rootfs}-GPT
+	if [ -n "$1" ];then
+		board=$board-$1
+	fi
+
+	echo -e "File name is \e[36m $board \e[0m"
+	read -t 10 -e -p "Rename the file? [N|y]" ANS 
+	ANS=${ANS:-n}
+	
+	case $ANS in
+			Y|y|yes|YES|Yes) rename=1;;
+			N|n|no|NO|No) rename=0;;
+			*) rename=0;;
+	esac
+	if [[ ${rename} == "1" ]]; then
+		read -e -p "Enter new file name: " IMGNAME
+		IMGNAME=$IMGNAME
+	fi
+	IMGNAME=${IMGNAME:-$board}
+	IMGNAME=${IMGNAME}-${day}-${time}.img
+}
+
+
+function build_sdbootimg(){
+	packm="unpack"
+	[[ -n "$1" ]] && [[ $1 != "-p" ]] && usage 
+	[[ -n "$1" ]] && packm="pack"
+
+	gen_file_name SDBOOT
+
+	if [ $packm == "pack" ];then
+		cd rockdev && ./version.sh $IMGNAME init && cd -
+	fi
+
+	IMAGE_PATH=$TOP_DIR/rockdev
+	PACK_TOOL_DIR=$TOP_DIR/tools/linux/Linux_Pack_Firmware
+
+	echo "Make sdboot.img"
+	cd $PACK_TOOL_DIR/rockdev && ./mksdbootimg.sh && cd -
+	mv $PACK_TOOL_DIR/rockdev/sdboot.img $IMAGE_PATH/pack/$IMGNAME
+	if [ $? -eq 0 ]; then
+		echo "Make sdboot image ok!"
+		echo -e "\e[36m $IMAGE_PATH/pack/$IMGNAME \e[0m"
+	else
+	   echo "Make sdboot image failed!"
+	   exit 1
+	fi
+	if [ $packm == "pack" ];then
+		cd rockdev && ./version.sh $IMGNAME pack && cd -
+	fi
+}
+
 function build_updateimg(){
+	packm="unpack"
+	[[ -n "$1" ]] && [[ $1 != "-p" ]] && usage 
+	[[ -n "$1" ]] && packm="pack"
+
+	gen_file_name 
+
+	if [ $packm == "pack" ];then
+		cd rockdev && ./version.sh $IMGNAME init $2 && cd -
+	fi
+
 	IMAGE_PATH=$TOP_DIR/rockdev
 	PACK_TOOL_DIR=$TOP_DIR/tools/linux/Linux_Pack_Firmware
     if [ "$RK_LINUX_AB_ENABLE"x = "true"x ];then
@@ -236,16 +348,77 @@ function build_updateimg(){
         fi
 
     else
-        echo "Make update.img"
-        cd $PACK_TOOL_DIR/rockdev && ./mkupdate.sh && cd -
-        mv $PACK_TOOL_DIR/rockdev/update.img $IMAGE_PATH
-        if [ $? -eq 0 ]; then
-            echo "Make update image ok!"
-        else
-            echo "Make update image failed!"
-            exit 1
-        fi
+		echo "Make update.img"
+		cd $PACK_TOOL_DIR/rockdev && ./mkupdate.sh && cd -
+		mv $PACK_TOOL_DIR/rockdev/update.img $IMAGE_PATH/pack/$IMGNAME
+		rm -rf $IMAGE_PATH/update.img
+		if [ $? -eq 0 ]; then
+		   echo "Make update image ok!"
+		   echo -e "\e[36m $IMAGE_PATH/pack/$IMGNAME \e[0m"
+		else
+		   echo "Make update image failed!"
+		   exit 1
+		fi
+
+		if [ $packm == "pack" ];then
+			cd rockdev && ./version.sh $IMGNAME pack && cd -
+		fi
     fi
+}
+
+function build_sdupdateimg(){
+
+	gen_file_name sdupdate
+
+	IMAGE_PATH=$TOP_DIR/rockdev
+	PACK_TOOL_DIR=$TOP_DIR/tools/linux/Linux_Pack_Firmware
+
+	echo "Make sdupdate.img"
+	if [ -f $SD_PARAMETER ]
+	then
+		echo -n "create parameter..."
+		ln -s -f $SD_PARAMETER $ROCKDEV/parameter.txt
+		echo "done."
+	else
+		echo -e "\e[31m error: $SD_PARAMETER not found! \e[0m"
+		exit 1
+	fi
+
+	if [[ x"$RK_SD_PACKAGE_FILE" != x ]];then
+		RK_PACK_TOOL_DIR=$TOP_DIR/tools/linux/Linux_Pack_Firmware/rockdev/
+		cd $RK_PACK_TOOL_DIR
+		rm -f package-file
+		ln -sf $RK_SD_PACKAGE_FILE package-file
+	fi
+
+	MKSDUPDATE_FILE=${RK_TARGET_PRODUCT}-mksdupdate.sh
+	if [[ x"$MKSDUPDATE_FILE" != x-mksdupdate.sh ]];then
+		rm -f mksdupdate.sh
+		ln -s $MKSDUPDATE_FILE mksdupdate.sh
+	fi
+
+	cd $PACK_TOOL_DIR/rockdev && ./mksdupdate.sh && cd -
+	mv $PACK_TOOL_DIR/rockdev/sdupdate.img $IMAGE_PATH/pack/$IMGNAME
+	rm -rf $IMAGE_PATH/sdupdate.img
+
+	if [ $? -eq 0 ]; then
+	   echo "Make sdupdate image ok!"
+	   echo -e "\e[36m $IMAGE_PATH/pack/$IMGNAME \e[0m"
+	else
+	   echo "Make sdupdate image failed!"
+	fi
+
+	if [ -f $PARAMETER ]
+	then
+		ln -s -f $PARAMETER $ROCKDEV/parameter.txt
+	fi
+
+	if [[ x"$RK_PACKAGE_FILE" != x ]];then
+		RK_PACK_TOOL_DIR=$TOP_DIR/tools/linux/Linux_Pack_Firmware/rockdev/
+		cd $RK_PACK_TOOL_DIR
+		rm -f package-file
+		ln -sf $RK_PACKAGE_FILE package-file
+	fi
 }
 
 function build_ota_ab_updateimg(){
@@ -307,6 +480,9 @@ if [ $BUILD_TARGET == uboot ];then
 elif [ $BUILD_TARGET == kernel ];then
     build_kernel
     exit 0
+elif [ $BUILD_TARGET == extboot ];then
+    build_extboot_image
+    exit 0
 elif [ $BUILD_TARGET == modules ];then
     build_modules
     exit 0
@@ -336,7 +512,13 @@ elif [ $BUILD_TARGET == debian ];then
     build_debian
     exit 0
 elif [ $BUILD_TARGET == updateimg ];then
-    build_updateimg
+    build_updateimg $2 $3
+    exit 0
+elif [ $BUILD_TARGET == sdbootimg ];then
+    build_sdbootimg $2
+    exit 0
+elif [ $BUILD_TARGET == sdupdateimg ];then
+    build_sdupdateimg
     exit 0
 elif [ $BUILD_TARGET == otapackage ];then
     build_ota_ab_updateimg
@@ -360,11 +542,34 @@ elif [ $BUILD_TARGET == allsave ];then
     build_all_save
     exit 0
 elif [ -f $NEW_BOARD_CONFIG ];then
+    if [ ! -n "$NEW_BOARD_CONFIG" ];then
+	    echo "==============================="
+	    echo "ERR:  $1 not found  "
+    	    echo "Can't found build config, please check again"
+	    echo "ls device/rockchip/rkxxxx"
+	    usage
+	    exit 1
+	fi
+    echo $NEW_BOARD_CONFIG
     rm -f $BOARD_CONFIG
     ln -s $NEW_BOARD_CONFIG $BOARD_CONFIG
-else
-    echo "Can't found build config, please check again"
-    usage
-    exit 1
+	unset RK_PACKAGE_FILE
+	unset RK_MKUPDATE_FILE
+	source $NEW_BOARD_CONFIG
+	if [[ x"$RK_PACKAGE_FILE" != x ]];then
+		PACK_TOOL_DIR=$TOP_DIR/tools/linux/Linux_Pack_Firmware/rockdev/
+        cd $PACK_TOOL_DIR
+		rm -f package-file
+        ln -sf $RK_PACKAGE_FILE package-file
+	fi
+    
+    MKUPDATE_FILE=${RK_TARGET_PRODUCT}-mkupdate.sh 
+    if [[ x"$MKUPDATE_FILE" != x-mkupdate.sh ]];then
+		PACK_TOOL_DIR=$TOP_DIR/tools/linux/Linux_Pack_Firmware/rockdev/
+        cd $PACK_TOOL_DIR
+        rm -f mkupdate.sh
+		ln -sf $MKUPDATE_FILE mkupdate.sh
+	fi
+    exit 0
 fi
